dataHandler: remove redundant markTask, unmarkTask and deleteTask. Instead, add updateList
